Silence Clang "mismatched-tags" warnings
Clang warns is declarations and definitions of class/struct do not
match. This is a somewhat spurious error as they are interchangable
(the warning appears to be for full MSVC compatibility).

"Fix" in fhicl-cpp by using consistent names as this only occurs in
one place.

Longer term, should review use of adding -Wno-mismatched-tags to
default Clang flags in line with GCC behaviour. Warning is otherwise
activated by -Wall.