Temporary full resolution of << operator
Clang (Apple LLVM version 7.0.2 (clang-700.1.81)) errors out
on compiling table_insertionOperator_t.cc due to ambiguity between
operator<< of the internal Logger struct and the free operator defined
in Table.h and OptionalTable.h. It is not yet clear if this is Clang
bug, or a strict C++ issue (that GNU allows).

In order to temporarily test fhiclcpp and allow an install, fully
resolve operator to the required one using ".operator<<". This
is not a fix, so problem may occur in deps such as messagefacility.